requirements.txt
PyPDF2>=3.0.1
pydantic>=2.0
python-dotenv
pathlib
typing-extensions
pymupdf>=1.23.0
nltk>=3.8
regex>=2023.0.0
langchain>=0.1.0
langchain-openai
langchain-community
faiss-cpu
openai
numpy
tiktoken